refactor: LimitedChessPiece 구현체의 코드 중복 삭제
