+++ src/main/java/domain/MoveStrategy.java
@@ -1,0 +1,4 @@
+package domain;
+
+public interface MoveStrategy {
+}
